Phase 5 + v0.1.0: CLI polish, docs site, release prep
cargo taut check (new): IR drift detection with baseline file
(taut/ir.snapshot.json). Bootstraps with --write; CI runs without --write
to fail the build on shape mismatch. Reports added/removed/changed
procedures and types in unified-diff style.

cargo taut inspect (new): human-readable IR rendering. --format table
(default) shows Procedures and Types tables; --format json pretty-prints
the IR; --format mermaid emits a flowchart LR diagram of procedures and
their input/output type references.

Bigint coercion fix in codegen: u64/i64/u128/i128 outputs now emit
v.pipe(v.union([v.bigint(), v.number(), v.string()]),
v.transform(BigInt), v.bigint()) (Valibot) and equivalent Zod, so
JSON-shipped numbers parse cleanly through output validation. Reverted
phase4-validate User.id back to u64.

SSE keep-alive restored (was disabled during axum 0.8 upgrade in Phase 4).
Sse::new(...).keep_alive(KeepAlive::default()) now wraps the response so
HTTP-aware proxies don't drop idle subscriptions. /rpc/_version endpoint
added returning {taut_rpc, ir_version}.

Macro errors now reference SPEC sections: every syn::Error from
#[rpc] / #[derive(Type)] / #[derive(Validate)] / #[derive(TautError)]
appends 'see SPEC §<N>' so users can find the rules.

#[derive(TautError)] variant attribute parsing now ignores foreign keys
(rename, optional, undefined, length, min, max, etc.) instead of
erroring. The three derives cooperate cleanly in the same #[taut(...)]
namespace.

taut_rpc::prelude module ships convenient re-exports for typical users:
use taut_rpc::prelude::*; brings rpc, Type, Validate, TautError, Router,
TautType, dump_if_requested into scope.

Documentation site (mdBook):
- New chapters: about, concepts/how-it-works, guides/troubleshooting,
  guides/migrate-from-rspc, guides/migrate-from-trpc, guides/deployment,
  guides/mcp, tutorials/{todo-app,realtime-chat,form-validation},
  reference/{cli,error-codes,faq,beyond-0.1,comparison,performance}.
- Refreshed: intro, architecture, type-mapping, wire-format,
  guides/getting-started, guides/validation.
- SUMMARY.md consolidates all entries.

Examples:
- New: examples/todo-react (axum + Vite + React full-stack with
  subscriptions and validation), examples/counter-sveltekit (axum +
  SvelteKit). Both port-numbered to avoid conflicts (7710, 7712).
- Bumped axum 0.7 -> 0.8, tower-http 0.5 -> 0.6, thiserror 1 -> 2 across
  all example servers to match the runtime crate's deps.
- examples/README.md catalogs every example.
- examples/phase1: AddError and GetUserError now derive TautError
  (required by Phase 4's macro-injected validation hook).

Release prep:
- All Cargo.toml versions bumped to 0.1.0; npm package.json to 0.1.0.
- Cargo metadata: keywords, categories, homepage, documentation,
  [package.metadata.docs.rs] block.
- Release workflow: .github/workflows/release.yml triggered on
  v*.*.* tags; verifies (test+clippy+fmt), publishes crates in
  dependency order, publishes npm, creates GitHub release with body
  extracted from CHANGELOG.
- Docs deploy: .github/workflows/docs.yml deploys mdBook to GitHub Pages
  on push to main.
- CODE_OF_CONDUCT.md (Contributor Covenant 2.1), SECURITY.md (private
  reporting + supported versions), refreshed CONTRIBUTING / issue and
  PR templates.
- CHANGELOG: [0.1.0] - 2026-05-07 entry with highlights, versioning
  policy, breaking changes from 0.0.0; pre-0.1 phase entries archived
  under 'Pre-0.1.0 development log'.
- ROADMAP: top banner pins v0.1.0 release; per-phase headings marked
  shipped; speculative items moved to 'Post-0.1.0 — under consideration'.
- README quality pass: badges (crates.io / npm / docs.rs / license / CI),
  install instructions, refreshed quick taste.
- .github/RELEASE_NOTES_v0.1.0.md drafts the release announcement.

Polish:
- All public API items in taut-rpc carry rustdoc comments
  (missing_docs warnings from 107 -> 0).
- 8 new trybuild fixtures plus dev-deps wired so rpc-stream-pass.rs
  compiles in the sandbox.
- StandardError catalog test pins every variant's code/status to the
  documented contract.
- impl Display for TypeRef and Primitive (Rust-syntax-flavoured;
  used by inspect's table format).
- scripts/smoke-all.sh runs every Phase 0-4 example end-to-end.
- cargo audit / cargo deny / npm audit: 0 vulnerabilities, 0 violations.

Tests: 417 passed, 0 failed, 1 ignored across the workspace +
trybuild suite (15/15 fixtures green).

cargo fmt --all clean. cargo check --workspace --all-targets clean.

Known v0.2 follow-ups noted in code:
- WebSocket transport (--features ws) needs Message::Text(Utf8Bytes)
  signature update for axum 0.8.
- Output JSON serialization: ProcedureResult::Err.code could become
  Cow<'static, str> to skip an allocation per error.
- Per-call validate override (currently per-client only).
- Async / cross-field validation (use #[taut(custom)] for now).

+++ examples/phase4-validate/server/src/main.rs
@@ -64,11 +64,12 @@ pub struct CreateUser {
 /// covers the recv path).
 #[derive(Serialize, Deserialize, Type)]
 pub struct User {
-    /// `u32` rather than `u64` to keep the demo simple — the codegen emits
-    /// `v.bigint()` for `u64` outputs, but the wire JSON ships them as plain
-    /// JS numbers, so validating on `bigint` fails on the recv path. v0.2
-    /// will add a coercion adapter; until then `u32` skirts the issue.
-    pub id: u32,
+    /// `u64` — the codegen now emits a coercion-style schema for native-bigint
+    /// primitives (union of bigint/number/string + transform → bigint), so the
+    /// wire JSON's plain JS number round-trips cleanly into a `bigint` at
+    /// output-validation time. The TS-side type is still `bigint`, matching
+    /// SPEC §3.1.
+    pub id: u64,
     pub username: String,
 }
